refactor: extract offer playlist skeleton in ui to reuse it later
